requirements.txt
Django == 5.0.3
Pillow == 10.2.0
djangorestframework == 3.15.0
django-filter == 24.2
psycopg2-binary == 2.9.9
python-dotenv == 1.0.1